LPS-56926 Move isValidTag from HtmlBBCodeTranslatorImpl to BBCodeParser.
